Extract Hatch discovery and clarify parsed settings
